polish: address deferred review minors
- DEFAULT_P: u32 -> u8 (removes the as-u8 truncation footgun)
- add Error::Serialize variant; to_cbor maps to it instead of misusing Error::Crypto
- totp_code: debug_assert!(digits <= 9) so 10^digits fits in u32
- hibp: hoist the function-local use to module scope in cmd_check
- generator: add succeeds_at_minimum_length boundary test (now 37 tests)

+++ src/crypto.rs
@@ -19,7 +19,7 @@ pub struct KdfParams {
 impl KdfParams {
     pub const DEFAULT_M: u32 = 65536; // 64 MiB
     pub const DEFAULT_T: u32 = 3;
-    pub const DEFAULT_P: u32 = 4;
+    pub const DEFAULT_P: u8 = 4;
 
     /// Fresh default-strength parameters with a new random salt.
     pub fn generate_default() -> Self {
@@ -28,7 +28,7 @@ impl KdfParams {
         KdfParams {
             m_cost: Self::DEFAULT_M,
             t_cost: Self::DEFAULT_T,
-            p_cost: Self::DEFAULT_P as u8,
+            p_cost: Self::DEFAULT_P,
             salt,
         }
     }
@@ -37,7 +37,7 @@ impl KdfParams {
     pub fn is_weaker_than_default(&self) -> bool {
         self.m_cost < Self::DEFAULT_M
             || self.t_cost < Self::DEFAULT_T
-            || (self.p_cost as u32) < Self::DEFAULT_P
+            || self.p_cost < Self::DEFAULT_P
     }
 }
